# Yul: Avoid expensive operations in functions called very often