--- a/libyul/optimiser/SyntacticalEquality.cpp
+++ b/libyul/optimiser/SyntacticalEquality.cpp
@@ -64,9 +64,8 @@ bool SyntacticallyEqual::expressionEqual(Identifier const& _lhs, Identifier cons
 }
 bool SyntacticallyEqual::expressionEqual(Literal const& _lhs, Literal const& _rhs)
 {
-	yulAssert(validLiteral(_lhs), "Invalid lhs literal during syntactical equality check");
-	yulAssert(validLiteral(_rhs), "Invalid rhs literal during syntactical equality check");
-
+	assert(validLiteral(_lhs));
+	assert(validLiteral(_rhs));
 	return _lhs.value == _rhs.value;
 }