Two changes are proposed here:

1. Do not include expensive assertion check in the equality operator.
2. Avoid expansive regex match (`std::regex`) whenever possible.

Add 1: The equality comparison is called very often in `CommonSubexpressionEliminator`, presumably because it is the comparison operator used for the map `CommonSubexpressionEliminator::m_replacementCandidates`. The method `validLiteral` is nontrivial and should not be called (twice!) on every call to the equality comparison.
It might be desirable to re-introduce the check in a more suitable place.

Add 2: Method `Parser::fetchDebugDataFromComment` was matching against a regex, but most of the time it was trying to match against an empty string. `std::regex` is known to be extremely slow. This would not be such a problem if this method was not called after every parsed token, see `Parser::advance`.
I believe even better solution would be to move the call to `fetchDebugDataFromComment` to a more appropriate place; my intuition says it is currently misplaced.