@blishko 2024-11-15
Benchmarking on my local laptop shows consistent improvement on external benchmarks:

Before:
|         File          | Pipeline |   Time   | Memory (peak) | Exit code |
|-----------------------|----------|---------:|--------------:|----------:|
| openzeppelin-5.0.2    |       ir |     14 s |       465 MiB |         0 |
| openzeppelin-4.9.0    |       ir |     18 s |       573 MiB |         0 |
| liquity-2024-10-30    |       ir |     19 s |       802 MiB |         0 |
| openzeppelin-4.7.0    |       ir |     26 s |       592 MiB |         0 |
| openzeppelin-4.8.0    |       ir |     29 s |       692 MiB |         0 |
| uniswap-v4-2024-06-06 |       ir |     53 s |      1533 MiB |         0 |
| eigenlayer-0.3.0      |       ir |    221 s |      3059 MiB |         0 |
| sablier-v2-1.2.0      |       ir |    627 s |      6192 MiB |         0 |


After:
|         File          | Pipeline |   Time   | Memory (peak) | Exit code |
|-----------------------|----------|---------:|--------------:|----------:|
| openzeppelin-5.0.2    |       ir |     14 s |       498 MiB |         0 |
| openzeppelin-4.9.0    |       ir |     18 s |       607 MiB |         0 |
| liquity-2024-10-30    |       ir |     18 s |       812 MiB |         0 |
| openzeppelin-4.7.0    |       ir |     24 s |       595 MiB |         0 |
| openzeppelin-4.8.0    |       ir |     28 s |       715 MiB |         0 |
| uniswap-v4-2024-06-06 |       ir |     50 s |      1479 MiB |         0 |
| eigenlayer-0.3.0      |       ir |    211 s |      3167 MiB |         0 |
| sablier-v2-1.2.0      |       ir |    592 s |      6312 MiB |         0 |