@blishko 2024-11-20
I don't have benchmarking results for the changes separately, although I could compute them if necessary.
Regarding the first point: We discussed this with @cameel. `assert` is called only in Debug mode. In Release mode, the macro results in no-op, so the calls to `validLiteral` disappear. `yulAssert` is always called. So with the proposed change we still get the sanity checks in Debug mode, but in Release mode, that is, in production, the compiler will no longer suffer from the performance penalty.