@cameel 2024-11-21
> I am not sure about the changes related to Add 1. Do you have the benchmarks with only `Add 2`?

This is what it looks like for me:

### IR
|         File          | Time (before) | Time (commit 1) | Time (after) |
|-----------------------|--------------:|----------------:|-------------:|
| uniswap-v4-2022-06-16 |          17 s |            19 s |         18 s |
| openzeppelin-5.0.2    |          28 s |            27 s |         28 s |
| openzeppelin-4.9.0    |          36 s |            35 s |         36 s |
| liquity-2024-10-30    |          39 s |            38 s |         43 s |
| openzeppelin-4.7.0    |          51 s |            50 s |         52 s |
| openzeppelin-4.8.0    |          56 s |            54 s |         57 s |
| uniswap-v4-2024-06-06 |         112 s |           105 s |        118 s |
| eigenlayer-0.3.0      |         497 s |           479 s |        511 s |
| sablier-v2-1.2.0      |        1378 s |          1371 s |       1347 s |

### Legacy
|         File          | Time (before) | Time (commit 1) | Time (after) |
|-----------------------|--------------:|----------------:|-------------:|
| uniswap-v4-2022-06-16 |           5 s |             5 s |          5 s |
| openzeppelin-5.0.2    |           9 s |             9 s |         10 s |
| openzeppelin-4.9.0    |          11 s |            12 s |         12 s |
| liquity-2024-10-30    |          10 s |             9 s |         10 s |
| openzeppelin-4.7.0    |          17 s |            17 s |         19 s |
| openzeppelin-4.8.0    |          20 s |            20 s |         21 s |
| uniswap-v4-2024-06-06 |          20 s |            20 s |         22 s |
| eigenlayer-0.3.0      |          75 s |            70 s |         78 s |
| sablier-v2-1.2.0      |         199 s |           212 s |        209 s |