@blishko 2024-11-21
> > I am not sure about the changes related to Add 1. Do you have the benchmarks with only `Add 2`?
> 
> This is what it looks like for me:
> 

Oh, this almost suggests we should revert the second commit.
And I am buffled how the first commit could have made sablier-v2-1.2.0 on legacy slower, because we are just removing an operation. Moreover on a code path that is not executed in legacy mode, no? That seems to be a fluke?